refactor(#9): 컴포넌트 파일 분리 및 스타일드 컴포넌트 재사용성 개선
