+++ src/components/Common/PostCardComponents.js
@@ -1,37 +1,4 @@
 import styled from "styled-components";
-import eyeIcon from "../../assets/images/eye.png";
-import heartIcon from "../../assets/images/heart.png";
-import commentIcon from "../../assets/images/comment.png";
-
-const InteractionsWrapper = styled.div`
-  display: flex;
-`;
-
-const InteractionItemWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  margin-right: 8px;
-`;
-
-const Icon = styled.img`
-  width: 16px;
-  height: 16px;
-  margin-right: 4px;
-`;
-
-const IconText = styled.span`
-  font-size: 13px;
-`;
-
-const DefaultArrow = styled.button`
-  width: 40px;
-  height: 24px;
-  border-radius: 16px 0 0 16px;
-`;
-
-const MirroredArrow = styled(DefaultArrow)`
-  transform: scaleX(-1);
-`;
 
 const ContentContainer = styled.div`
   flex-grow: 1;
@@ -40,6 +7,11 @@ const ContentContainer = styled.div`
   border-bottom: 1px solid;
   padding: 10px 0;
   margin: 10px 0 10px 0;
+`;
+
+const ThumbnailWrapper = styled.div`
+  flex-grow: 1;
+  border: 1px solid;
 `;
 
 const ProfileWrapper = styled.div`
@@ -61,28 +33,6 @@ const ProfileInfo = styled.div`
   flex-direction: column;
 `;
 
-const InteractionItem = ({ icon, count }) => (
-  <InteractionItemWrapper>
-    <Icon src={icon} />
-    <IconText>{count}</IconText>
-  </InteractionItemWrapper>
-);
-
-export const Interaction = ({ count }) => (
-  <InteractionsWrapper>
-    <InteractionItem icon={eyeIcon} count={count.view} />
-    <InteractionItem icon={heartIcon} count={count.like} />
-    <InteractionItem icon={commentIcon} count={count.comment} />
-  </InteractionsWrapper>
-);
-
-export const ArrowButton = () => (
-  <div>
-    <DefaultArrow></DefaultArrow>
-    <MirroredArrow></MirroredArrow>
-  </div>
-);
-
 export const PostProfile = ({ name, job }) => (
   <ProfileWrapper>
     <ProfileImage />
@@ -99,3 +49,5 @@ export const PostContent = ({ title, content }) => (
     <p>{content}</p>
   </ContentContainer>
 );
+
+export const Thumbnail = () => <ThumbnailWrapper></ThumbnailWrapper>;
